yubioath cli: fix handling of otpauth URIs
If we use the 'or's here we incorrectly pull a bunch of values from the
command-line arguments instead of from the parsed URL. Even if the user
doesn't specify these values (e.g. digits) on the command line, they
already have Click-defined default values. So regardless of the 'digits'
value in the otpauth:// URI, '6' would always be used.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>