Partially addresses #187
This PR doesn't address the whole issue, but it does add the up/down
navigation when the cursor is in the search bar. I don't know anything
about qml, so this is probably not clean by any means, but I'm happy to
change it. Maybe I'll learn more about how to write this stuff the
proper way as I continue to chip away at the requests outlined in #187.