Include above-and-left-of-primary pixels in QR screenshots
`QGuiApplication::primaryScreen()->grabWindow(...)` only includes pixels
below and to the right of the top left corner of the primary screen (at
least in Linux). Consider a screen arrangement like:

                          +-------------+
    +---------++---------+|             |
    +         ||         ||             |
    +    3    || PRIMARY ||      2      |
    +---------++---------++-------------+

With this screen arrangement, the top of screen 2 and all of screen 3
would be left out of the screenshot.

Combining the results of calling `grabWindow(...)` on each `QScreen`
resolves the issue.