Sort qrc input file list
so that yubioath-desktop packages build in a reproducible way
in spite of indeterministic filesystem readdir order

See https://reproducible-builds.org/ for why this is good.

Note: With the current pretty flat resource layout,
dirs has at most 1 entry so would not need to be sorted,
but OTOH sorting this will be very fast.