Catch exception in signal.signal()
In Python 3.8, `signal.signal()` raises an exception when called on any
thread other than the main thread.

Catching the exception means the program can resume as normal - except
this means the signal handler won't be set up, so we should probably try
to move this to the C++/QML layer instead.