Avoid moving focus outside of material app
Before this change the GlobalShortcuts was wrapping the material app.
As a consequence triggering the escape intent a few times could move the
focus outside of the material app, making navigation through focus
traversal not working. This commit changes the position of the
GlobalShortcuts in the widget tree. It also ensures that when the escape
intent is triggered we never move focus outside the material app.