Linux: fix the tray icon when running in a sandbox
Pass only the icon name rather than the path when running in a flatpak/snap sandbox
since the path will not be visible in the host environment.

The tray_manager PR https://github.com/leanflutter/tray_manager/pull/43 already has
the merged change that skips adding full path when running in a sandbox.