Handle region code in `update_android_strings.py`
The `update_android_strings.py`script  will now create the correct source
directory for languages with region code. However, manual work is still
required if the language with the region does not have translations for
the android strings in the corresponding arb file. In that case we need
to manually copy the content from the language without the region
code.